+++ src/core/createBoundingBox.js
